+++ Functions/Assertions/ContainExactly.ps1
@@ -1,12 +1,12 @@
-function PesterContainExactly($file, $contentExpecation) {
-    return ((& $SafeCommands['Get-Content'] -Encoding UTF8 $file) -cmatch $contentExpecation)
+function PesterContainExactly($file, $contentExpectation) {
+    return ((& $SafeCommands['Get-Content'] -Encoding UTF8 $file) -cmatch $contentExpectation)
 }
 
-function PesterContainExactlyFailureMessage($file, $contentExpecation) {
-    return "Expected: file ${file} to contain exactly {$contentExpecation}"
+function PesterContainExactlyFailureMessage($file, $contentExpectation) {
+    return "Expected: file ${file} to contain exactly {$contentExpectation}"
 }
 
-function NotPesterContainExactlyFailureMessage($file, $contentExpecation) {
-    return "Expected: file {$file} to not contain exactly ${contentExpecation} but it did"
+function NotPesterContainExactlyFailureMessage($file, $contentExpectation) {
+    return "Expected: file {$file} to not contain exactly ${contentExpectation} but it did"
 }